+++ css/csshome.css
@@ -1,0 +1,288 @@
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+  text-decoration: none;
+}
+header {
+  background: lightblue;
+  padding-bottom: 100px;
+  color: white;
+}
+body {
+  margin-top: 0px;
+  margin-left: 10px;
+  padding: 0px;
+  background: url("bg2.jpg");
+  height: 840px;
+}
+.left {
+  color: red;
+  display: inline;
+  font-size: 25px;
+}
+.left span {
+  border: 2px solid black;
+  background-color: white;
+  padding: 1px;
+}
+.left img {
+  border: 2px solid black;
+  border-radius: 2px;
+  display: inline;
+  width: 145px;
+  height: 90px;
+  margin-left: 10px;
+  margin-top: 10px;
+}
+.mid {
+  display: inline;
+  border: 2px solid black;
+  background-color: white;
+  right: 190px;
+  border-radius: 2px;
+  color: red;
+  padding-top: 5px;
+  padding-bottom: 5px;
+  font-size: 24px;
+  top: 20px;
+}
+.mid a {
+  color: red;
+  padding: 10px;
+  text-decoration: none;
+}
+.right a {
+  color: red;
+  padding: 1px;
+  text-decoration: none;
+}
+.right {
+  background-color: white;
+  display: inline;
+  border: 2px solid black;
+  position: absolute;
+  right: 5px;
+  font-size: 25px;
+  top: 20px;
+}
+.headca {
+  left: 160px;
+  top: 5px;
+}
+.about,
+.contact,
+footer {
+  font-family: "Poppins", sans-serif;
+}
+.contact .contact-content {
+  display: flex;
+  flex-wrap: wrap;
+  align-items: center;
+  justify-content: space-between;
+}
+.navbar {
+  position: absolute;
+  width: 100%;
+  z-index: 999;
+  padding: 30px 30px;
+  font-family: "Ubuntu", sans-serif;
+  transition: all 0.3s ease;
+  background-color: rgb(54, 10, 104);
+  margin-top: 0px;
+  margin-left: 0px;
+  margin-right: 0px;
+}
+.navbar.sticky {
+  padding: 15px 0;
+  background: rgb(20, 190, 220);
+}
+.navbar .max-width {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+.navbar .logo a {
+  color: #fff;
+  font-size: 35px;
+  font-weight: 600;
+}
+.navbar .logo a span {
+  color: rgb(20, 190, 220);
+  transition: all 0.3s ease;
+}
+.navbar.sticky .logo a span {
+  color: #fff;
+}
+.navbar .menu li {
+  list-style: none;
+  display: inline-block;
+}
+.navbar .menu li a {
+  display: block;
+  color: #fff;
+  font-size: 18px;
+  font-weight: 500;
+  margin-left: 25px;
+  transition: color 0.3s ease;
+}
+.navbar .menu li a:hover {
+  color: rgb(20, 190, 220);
+}
+.navbar.sticky .menu li a:hover {
+  color: #fff;
+}
+/* menu btn styling */
+.menu-btn {
+  color: #fff;
+  font-size: 23px;
+  cursor: pointer;
+  display: none;
+  padding: 10px;
+}
+.footer {
+  position: fixed;
+  left: 0;
+  bottom: 0;
+  width: 100%;
+  background-color: rgba(250, 250, 250, 0.5);
+  color: black;
+  text-align: center;
+  font-size: 20px;
+}
+.footer a {
+  color: black;
+  text-decoration: none;
+}
+
+.home {
+  display: flex;
+  background: url("images/banner3.jpg") no-repeat center;
+  height: 100vh;
+  color: #fff;
+  min-height: 500px;
+  background-size: cover;
+  background-attachment: fixed;
+  font-family: "Ubuntu", sans-serif;
+}
+.home .max-width {
+  width: 100%;
+  display: flex;
+}
+.home .max-width .row {
+  margin-right: 0;
+}
+.home .home-content .text-1 {
+  font-size: 27px;
+}
+.home .home-content .text-2 {
+  font-size: 25px;
+  font-weight: 600;
+  margin-left: -3px;
+  margin-top: 160px;
+  margin-left: 20px;
+}
+.home .home-content .text-3 {
+  font-size: 25px;
+  margin-top: 160px;
+  margin-left: -3px;
+}
+.home .home-content .text-3 span {
+  color: aqua;
+  font-weight: 500;
+}
+.home .home-content a {
+  display: inline-block;
+  background: rgb(100, 200, 0);
+  color: #ff2b4a;
+  font-size: 25px;
+  padding: 12px 36px;
+  margin-top: 20px;
+  font-weight: 400;
+  border-radius: 6px;
+  border: 2px solid rgb(100, 200, 0);
+  transition: all 0.3s ease;
+}
+.home .home-content .text-2 .topic {
+  color: #15a1ff;
+  font-weight: 1000;
+  font-size: 40px;
+  margin-right: 10px;
+}
+
+home .home-content .note {
+  font-size: 10px;
+}
+
+.home .home-content .text-3 .topic {
+  color: #15a1ff;
+  font-size: 40px;
+  font-weight: 1000;
+  margin-right: 10px;
+}
+.home .home-content .text-3 .bold1 {
+  font-weight: bold;
+  color: #00ff18;
+  text-decoration: underline;
+  margin-right: 20px;
+}
+
+.home .home-content {
+  font-size: 60px;
+  color: #00ff18;
+  background-color: rgba(150, 0, 40, 0.5);
+  z-index: 800;
+  margin-right: 20px;
+}
+.home .home-content .text-2 img {
+  border: 20px solid black;
+  border-radius: 10px;
+
+  margin-right: 20px;
+}
+.home .home-content .text-3 .form {
+  border: 10px solid #15a1ff;
+  border-radius: 5px;
+  margin-right: 10px;
+  width: 300px;
+
+  margin: 20px;
+}
+
+.home .home-content .text-3 .form div {
+  font-size: 50px;
+}
+
+.home .home-content .text-3 .form {
+  background-color: White;
+  width: 600px;
+}
+
+.home .home-content .text-3 .form input,
+textarea {
+  background-color: #00ff2d;
+  margin-right: 10px;
+}
+
+.home .home-content .text-3 .form div textarea {
+  width: 400px;
+  height: 200px;
+}
+home .home-content .text-3 .form div input {
+  height: 200px;
+  height: 300px;
+}
+.home .home-content .text-3 .form div {
+  margin: 10px;
+  margin-right: 20px;
+  color: #15a1ff;
+}
+
+.footer .fab fa-github {
+  background: url("");
+}
+
+.footer .far fa-copyright {
+  font-size: 20px;
+  margin-right: 20px;
+}
