--- donat.htm
@@ -1,102 +1,0 @@
-<!DOCTYPE html>
-<html lang="en"> 
-				<head>	
-								<title>COVIDHELP</title>
-	<link rel="stylesheet" href="csshome.css">	<meta charset="UTF-8"> 
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-		<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.2/css/all.min.css"/> 
-		
-		</head> 
-		<body>	
-						<nav class="navbar">	
-											<div class="max-width">	
-															
-		<div class="logo"><a href="#">Covid<span> Help</span></a></div>	
-				<ul class="menu">		
-<li><a href="home.html" class="menu-btn">Home</a></li>		
-<li><a href="meter.html"				target="_blank"		class="menu-btn">Covid Count</a></li><li><a href="staysafe.htm" class="menu-btn">Why stay safe</a></li>	
-<li><a href="howstaysafe.htm" class="menu-btn">How stay safe</a></li>		
-<li><a href="vaccine.htm" class="menu-btn">Vaccine</a></li>
-<li><a href="doc.htm" class="menu-btn">Online Doctors</a></li>
-<li><a href="donat.htm" class="menu-btn">Beta Donation Programme</a></li> 	
-<li><a href="contactus.htm" class="menu-btn">Contact Us</a></li> 			</ul>
-<div class="menu-btn">				
-<i class="fas fa-bars"></i>		
-	</div>		</div>	</nav>	 
-	
-	
-	<section class="home" id="home"> 
-					<div class="max-width"> <div class="row"> <div class="home-content"> 
-									<div class="text-2"><span class="topic">What is Beta Donation Programmes:</span><br><br>It connects all NGOs over India.Many people are in great loss due to this pandemic.<br>
-But people can help each other through donations.<br><br>By this programme you just have to fill a form and<br> NGOs will arrive at your given address to take the item you want to donate for people who are in need.
-													
-													</div> 
-									<div class="text-3">
-													<br><br><hr><br><br>
-		<span class="topic">
-
-Fill the form below to apply for donation:
-
-  </span>	<br>NGOs will reach your location witinh 5-10		days to recieve your donation				
-<br><br><br>
-<form class="form">
-				
-			<div>Name: <br><input type="text" value="Enter your name"></div>
-			<br>
-			<div>Phone No. :<br><input type="text" value="Enter your Phone No."></div>
-			<br>
-			
-			
-				<div>
-								
-								Address:<br><textarea name="feedback" cols=20 row=10></textarea>
-								
-				</div>
-				<br>
-			<div>
-							When You want us to pick your donations:
-							<br>Date:<input type="date"	>
-							
-								
-								
-				</div>
-			<br>
-				<div>
-								
-								Description of Donation:<br><textarea name="feedback" cols=30 row=20></textarea>
-								
-				</div>
-				<br>
-				<div>
-							<input type="submit"	value="Submit Now">
-								
-				</div>
-			
-					<div>
-							<input type="reset"	value="Reset Now">
-								
-				</div>
-			
-			
-			
-			
-			
-			
-			</form>		
-			
-			<br><br><span><a href="places.htm">Places where pickup is available</a></span>				<br><br><br><br><br>			
-									<span class="typing"></span></div>
-									  </div> </div> </div>
-	 </section>
-	
-				<footer class="footer">
-						 <span >Created By <i class="fab fa-github"></i> <a href="https://github.com/Whitedevil4648/INIT-2022-COVID-SITE/blob/main/README.md">Anand Kumar(Discord-White Devil#4648)</a> | 
-						 <span class="far fa-copyright"></span> 2021 All rights reserved.</span> </footer>
-
- </body> 
-						 
-						 
-						 
-						 
-						 
-</html>
