+++ home.html
@@ -1,55 +1,73 @@
 <!DOCTYPE html>
-<html lang="en"> 
-				<head>	
-								<title>COVIDHELP</title>
-	<link rel="stylesheet" href="csshome.css">	<meta charset="UTF-8"> 
+<html lang="en">
+
+<head>
+	<title>COVIDHELP</title>
+	<link rel="stylesheet" href="css/csshome.css">
+	<meta charset="UTF-8">
 	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-		<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.2/css/all.min.css"/> 
-		
-		</head> 
-		<body>	
-						<nav class="navbar">	
-											<div class="max-width">	
-															
-		<div class="logo"><a href="#">Covid<span> Help</span></a></div>	
-				<ul class="menu">		
-<li><a href="home.html" class="menu-btn">Home</a></li>		
-<li><a href="meter.html"				target="_blank"		class="menu-btn">Covid Count</a></li><li><a href="staysafe.htm" class="menu-btn">Why stay safe</a></li>	
-<li><a href="howstaysafe.htm" class="menu-btn">How stay safe</a></li>		
-<li><a href="vaccine.htm" class="menu-btn">Vaccine</a></li>
-<li><a href="doc.htm" class="menu-btn">Online Doctors</a></li>
-<li><a href="donat.htm" class="menu-btn">Beta Donation Programme</a></li> 	
-<li><a href="contactus.htm" class="menu-btn">Contact Us</a></li> 			</ul>
-<div class="menu-btn">				
-<i class="fas fa-bars"></i>		
-	</div>		</div>	</nav>	 
-	
-	
-	<section class="home" id="home"> 
-					<div class="max-width"> <div class="row"> <div class="home-content"> 
-									<div class="text-2"><span class="topic">COVID had a darastic effect on our life.<br>But we have to find our own way to be happy and safe
-	<img src="P1.jpg" alt=""></div> 
-									<div class="text-3">This website will tell you all about how to stay safe.<br>Why should you stay safe.<br>What to do after infection. <br>Read all articles full information.<br><br><hr><br><br>
-		<span class="topic">ONLY WE CAN SAVE EACH OTHER,<br>SO STAY SAFE STAY HEALTHY
+	<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.2/css/all.min.css" />
+
+</head>
+
+<body>
+	<nav class="navbar">
+		<div class="max-width">
+
+			<div class="logo"><a href="#">Covid<span> Help</span></a></div>
+			<ul class="menu">
+				<li><a href="home.html" class="menu-btn">Home</a></li>
+				<li><a href="meter.html" target="_blank" class="menu-btn">Covid Count</a></li>
+				<li><a href="staysafe.htm" class="menu-btn">Why stay safe</a></li>
+				<li><a href="howstaysafe.htm" class="menu-btn">How stay safe</a></li>
+				<li><a href="vaccine.htm" class="menu-btn">Vaccine</a></li>
+				<li><a href="doc.htm" class="menu-btn">Online Doctors</a></li>
+				<li><a href="donat.htm" class="menu-btn">Beta Donation Programme</a></li>
+				<li><a href="contactus.htm" class="menu-btn">Contact Us</a></li>
+			</ul>
+			<div class="menu-btn">
+				<i class="fas fa-bars"></i>
+			</div>
+		</div>
+	</nav>
+
+
+	<section class="home" id="home">
+		<div class="max-width">
+			<div class="row">
+				<div class="home-content">
+					<div class="text-2"><span class="topic">COVID had a darastic effect on our life.<br>But we have to
+							find our own way to be happy and safe
+							<img src="P1.jpg" alt=""></div>
+					<div class="text-3">This website will tell you all about how to stay safe.<br>Why should you stay
+						safe.<br>What to do after infection. <br>Read all articles full information.<br><br>
+						<hr><br><br>
+						<span class="topic">ONLY WE CAN SAVE EACH OTHER,<br>SO STAY SAFE STAY HEALTHY
 
 
 
-  </span>							
-									
-									<span class="typing"></span></div>
-									  </div> </div> </div>
-	 </section>
-	
-		<footer class="footer">
-						 <span >Created By <i class="fab fa-github"></i> <a href="https://github.com/Whitedevil4648/INIT-2022-COVID-SITE/blob/main/README.md">Anand Kumar(Discord-White Devil#4648)</a> | 
-						 <span class="far fa-copyright"></span> 2021 All rights reserved.</span> </footer>
+						</span>
+
+						<span class="typing"></span>
+					</div>
+				</div>
+			</div>
+		</div>
+	</section>
+
+	<footer class="footer">
+		<span>Created By <i class="fab fa-github"></i> <a
+				href="https://github.com/Whitedevil4648/INIT-2022-COVID-SITE/blob/main/README.md">Anand
+				Kumar(Discord-White Devil#4648)</a> |
+			<span class="far fa-copyright"></span> 2021 All rights reserved.</span>
+	</footer>
 
 
 
- </body> 
-						 
-						 
-						 
-						 
-						 
+</body>
+
+
+
+
+
 </html>
