+++ js/back.js
@@ -1,0 +1,12 @@
+const http = require('http');
+
+const hostname = '127.0.0.1';
+const port = 3000;
+
+const server = http.createServer((req, res) => {
+  res.statusCode = 200;
+  res.setHeader('Content-Type', 'text/html');
+  res.end(index.html)});
+
+server.listen(port, hostname), () => {			
+  console.log('Server running at http://${hostname}:${port}/')};
