+++ meter.html
@@ -59,7 +59,7 @@ crossorigin = "anonymous"></script>
  }
  .container{
  	width: 90%;
- 	height: 155px
+ 	height: 155px;
  	margin:auto;
  }
 .total-case-box{
@@ -86,7 +86,7 @@ crossorigin = "anonymous"></script>
 .icon-box{
 	width: 50px;
 	height: 50px;
-	border-radius: 50%
+	border-radius: 50%;
 	background:#ff002f;
 	margin: auto;
 	display: flex;
