+++ staysafe.htm
@@ -1,71 +1,98 @@
 <!DOCTYPE html>
-<html lang="en"> 
-				<head>	
-								<title>COVIDHELP</title>
-	<link rel="stylesheet" href="csshome.css">	<meta charset="UTF-8"> 
+<html lang="en">
+
+<head>
+	<title>COVIDHELP</title>
+	<link rel="stylesheet" href="css/csshome.css">
+	<meta charset="UTF-8">
 	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-		<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.2/css/all.min.css"/> 
-		</head> 
-		<body>	
-						<nav class="navbar">	
-											<div class="max-width">	
-															
-		<div class="logo"><a href="#">Covid<span> Help</span></a></div>	
-				<ul class="menu">		
-<li><a href="home.html" class="menu-btn">Home</a></li>		
-<li><a href="meter.html"						class="menu-btn">Covid Count</a></li>	
-<li><a href="staysafe.htm" class="menu-btn">Why stay safe</a></li>	
-<li><a href="howstaysafe.htm" target="_blank" class="menu-btn">How stay safe</a></li>	
-	<li><a href="vaccine.htm" class="menu-btn">Vaccine</a></li>
-	
-	<li><a href="doc.htm" class="menu-btn">Online Doctors</a></li><li><a href="donat.htm" class="menu-btn">Beta Donation Programme</a></li>		<li><a href="contactus.htm" class="menu-btn">Contact Us</a></li> 			</ul>
-<div class="menu-btn">				<i class="fas fa-bars"></i>		
-	</div>		</div>	</nav>	 
-	
-						
-	<section class="home" id="home"> 
-					<div class="max-width"> <div class="row"> <div class="home-content"> 
-									<div class="text-2"><span class="topic">For may people the question is why should<br> one stay safe as for them virus is just a<br> fake news</span></div>
-									<div class="text-3"><span class="topic">How deadly is Corona Virus</span><br> <br>      
-	Here is a detailed explanation of how people die due to infection -:<br><br>
-1)First the Virus enters your body by nose,mouth,eyes,etc.<br><br>
-2)Human breathing system starts from nose and end in Alveoli(in lungs).Alveoli are site for exchange of gases(Oxygen and Carbon dioxide).<br><br>
-3)The virus somehow travel through your windpipe and enter the cells of wall of Alveoli.It upload its RNA into the cell's nucleus which make the cell generates more copies of that virus.At some point the cell burst releasing many copies of virus which goes into more cell and make then burst.<br><br>
-4)Bursting of cells makes pus which cover the site for exchange of gases.<br><br>
-5)Advancded immune system already starts  acting to produce antibodies.<br><br>
-6)Even if your immune system is strong,filling of pus in the alveoli becomes deadly.As it stops the exchange of gas and Oxygen level in your blood goes down.<br><br>
-7)You can now understand the symptoms-Cough to take the pus out of alveoli,Breathing problems due to lack of oxygen,fever because your body is fighting.<br><br>
-8)At last the person dies due to lack of Oxygen in his/her body.<br><br>
-<span class="topic">
-This is not for every case of Corona Virus,but what makes you so sure that your case will be normal with mild fever not the above case.</span> <br><br>
+	<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.2/css/all.min.css" />
+</head>
 
-Now i think you know why itis important to stay safe.<br><br>
+<body>
+	<nav class="navbar">
+		<div class="max-width">
 
-Here is detailed video explaination form YouTube watch it if you want to- <br>
-				
-<h1 class="topic">Corona Virus Demonstration</h1> 		<iframe width="90%" height="600px" src="https://www.youtube.com/embed/0G0vq3tVAKE" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>	<br><br><br>
-Now you should know how to stay safe.<br>
-Check the menu for next article on it.
+			<div class="logo"><a href="#">Covid<span> Help</span></a></div>
+			<ul class="menu">
+				<li><a href="home.html" class="menu-btn">Home</a></li>
+				<li><a href="meter.html" class="menu-btn">Covid Count</a></li>
+				<li><a href="staysafe.htm" class="menu-btn">Why stay safe</a></li>
+				<li><a href="howstaysafe.htm" target="_blank" class="menu-btn">How stay safe</a></li>
+				<li><a href="vaccine.htm" class="menu-btn">Vaccine</a></li>
+
+				<li><a href="doc.htm" class="menu-btn">Online Doctors</a></li>
+				<li><a href="donat.htm" class="menu-btn">Beta Donation Programme</a></li>
+				<li><a href="contactus.htm" class="menu-btn">Contact Us</a></li>
+			</ul>
+			<div class="menu-btn"> <i class="fas fa-bars"></i>
+			</div>
+		</div>
+	</nav>
 
 
+	<section class="home" id="home">
+		<div class="max-width">
+			<div class="row">
+				<div class="home-content">
+					<div class="text-2"><span class="topic">For may people the question is why should<br> one stay safe
+							as for them virus is just a<br> fake news</span></div>
+					<div class="text-3"><span class="topic">How deadly is Corona Virus</span><br> <br>
+						Here is a detailed explanation of how people die due to infection -:<br><br>
+						1)First the Virus enters your body by nose,mouth,eyes,etc.<br><br>
+						2)Human breathing system starts from nose and end in Alveoli(in lungs).Alveoli are site for
+						exchange of gases(Oxygen and Carbon dioxide).<br><br>
+						3)The virus somehow travel through your windpipe and enter the cells of wall of Alveoli.It
+						upload its RNA into the cell's nucleus which make the cell generates more copies of that
+						virus.At some point the cell burst releasing many copies of virus which goes into more cell and
+						make then burst.<br><br>
+						4)Bursting of cells makes pus which cover the site for exchange of gases.<br><br>
+						5)Advancded immune system already starts acting to produce antibodies.<br><br>
+						6)Even if your immune system is strong,filling of pus in the alveoli becomes deadly.As it stops
+						the exchange of gas and Oxygen level in your blood goes down.<br><br>
+						7)You can now understand the symptoms-Cough to take the pus out of alveoli,Breathing problems
+						due to lack of oxygen,fever because your body is fighting.<br><br>
+						8)At last the person dies due to lack of Oxygen in his/her body.<br><br>
+						<span class="topic">
+							This is not for every case of Corona Virus,but what makes you so sure that your case will be
+							normal with mild fever not the above case.</span> <br><br>
 
-<span class="typing"></span>
-									 </div>  </div> </div>
-	 </section>		
-				
-	
-	
-		<footer class="footer">
-						 <span >Created By <i class="fab fa-github"></i> <a href="https://github.com/Whitedevil4648/INIT-2022-COVID-SITE/blob/main/README.md">Anand Kumar(Discord-White Devil#4648)</a> | 
-						 <span class="far fa-copyright"></span> 2021 All rights reserved.</span> </footer>
-  
-						 
-						 
-						 
-						 
-						 
+						Now i think you know why itis important to stay safe.<br><br>
 
-		
-				
+						Here is detailed video explaination form YouTube watch it if you want to- <br>
+
+						<h1 class="topic">Corona Virus Demonstration</h1> <iframe width="90%" height="600px"
+							src="https://www.youtube.com/embed/0G0vq3tVAKE" title="YouTube video player" frameborder="0"
+							allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
+							allowfullscreen></iframe> <br><br><br>
+						Now you should know how to stay safe.<br>
+						Check the menu for next article on it.
+
+
+
+						<span class="typing"></span>
+					</div>
+				</div>
+			</div>
+	</section>
+
+
+
+	<footer class="footer">
+		<span>Created By <i class="fab fa-github"></i> <a
+				href="https://github.com/Whitedevil4648/INIT-2022-COVID-SITE/blob/main/README.md">Anand
+				Kumar(Discord-White Devil#4648)</a> |
+			<span class="far fa-copyright"></span> 2021 All rights reserved.</span>
+	</footer>
+
+
+
+
+
+
+
+
+
 </body>
+
 </html>
