+++ vaccine.htm
@@ -1,70 +1,85 @@
 <!DOCTYPE html>
-<html lang="en"> 
-				<head>	
-								<title>COVIDHELP</title>
-	<link rel="stylesheet" href="csshome.css">	<meta charset="UTF-8"> 
+<html lang="en">
+
+<head>
+	<title>COVIDHELP</title>
+	<link rel="stylesheet" href="css/csshome.css">
+	<meta charset="UTF-8">
 	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-		<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.2/css/all.min.css"/> 
-		</head> 
-		<body>	
-						<nav class="navbar">	
-											<div class="max-width">	
-															
-		<div class="logo"><a href="#">Covid<span> Help</span></a></div>	
-				<ul class="menu">		
-<li><a href="home.html" class="menu-btn">Home</a></li>		
-<li><a href="meter.html"						class="menu-btn">Covid Count</a></li>	
-<li><a href="staysafe.htm" class="menu-btn">Why stay safe</a></li>	
-<li><a href="howstaysafe.htm" target="_blank" class="menu-btn">How stay safe</a></li>	
-	<li><a href="vaccine.htm" class="menu-btn">Vaccine</a></li>
-	<li><a href="doc.htm" class="menu-btn">Online Doctors</a></li>
-	
-	<li><a href="donat.htm" class="menu-btn">Beta Donation Programme</a></li>
-		<li><a href="contactus.htm" class="menu-btn">Contact Us</a></li> 			</ul>
-<div class="menu-btn">				<i class="fas fa-bars"></i>		
-	</div>		</div>	</nav>	 
-	
-						
-	<section class="home" id="home"> 
-					<div class="max-width"> <div class="row"> <div class="home-content"> 
-									<div class="text-2"><span class="topic">   Vaccine</span></div>
-									<div class="text-3"><span class="topic">  Information on vaccine.</span><br> <br>   
-									Vaccination is very important as it provides you safety even after you are infected.
-							<br>		<br>There are rumours that the vaccines are having a side effect on human body,but its not true.<br><br>Get vaccinated as soon as possible.<br><br>
-<span class="topic">Website Portal for registration of Vaccine</span><br><br>
-1)<a href="https://www.cowin.gov.in/">Cowin Gov</a>
+	<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.2/css/all.min.css" />
+</head>
+
+<body>
+	<nav class="navbar">
+		<div class="max-width">
+
+			<div class="logo"><a href="#">Covid<span> Help</span></a></div>
+			<ul class="menu">
+				<li><a href="home.html" class="menu-btn">Home</a></li>
+				<li><a href="meter.html" class="menu-btn">Covid Count</a></li>
+				<li><a href="staysafe.htm" class="menu-btn">Why stay safe</a></li>
+				<li><a href="howstaysafe.htm" target="_blank" class="menu-btn">How stay safe</a></li>
+				<li><a href="vaccine.htm" class="menu-btn">Vaccine</a></li>
+				<li><a href="doc.htm" class="menu-btn">Online Doctors</a></li>
+
+				<li><a href="donat.htm" class="menu-btn">Beta Donation Programme</a></li>
+				<li><a href="contactus.htm" class="menu-btn">Contact Us</a></li>
+			</ul>
+			<div class="menu-btn"> <i class="fas fa-bars"></i>
+			</div>
+		</div>
+	</nav>
 
 
-<br><br><span class="topic">Changes in body after Vaccination:</span>
+	<section class="home" id="home">
+		<div class="max-width">
+			<div class="row">
+				<div class="home-content">
+					<div class="text-2"><span class="topic"> Vaccine</span></div>
+					<div class="text-3"><span class="topic"> Information on vaccine.</span><br> <br>
+						Vaccination is very important as it provides you safety even after you are infected.
+						<br> <br>There are rumours that the vaccines are having a side effect on human body,but its not
+						true.<br><br>Get vaccinated as soon as possible.<br><br>
+						<span class="topic">Website Portal for registration of Vaccine</span><br><br>
+						1)<a href="https://www.cowin.gov.in/">Cowin Gov</a>
 
-<br><br>1)Mild Headache
-<br><br>2)Mild Fever
-<br><br>3)Pain or Swelling at injections site
-<br><br>4)Irritation
-<br><br>These effects go away in few days after vaccination.
 
-	
- 
+						<br><br><span class="topic">Changes in body after Vaccination:</span>
+
+						<br><br>1)Mild Headache
+						<br><br>2)Mild Fever
+						<br><br>3)Pain or Swelling at injections site
+						<br><br>4)Irritation
+						<br><br>These effects go away in few days after vaccination.
 
 
 
-<span class="typing"></span>
-									 </div>  </div> </div>
-	 </section>		
-				
-	
-	
-		<footer class="footer">
-						 <span >Created By <i class="fab fa-github"></i> <a href="https://github.com/Whitedevil4648/INIT-2022-COVID-SITE/blob/main/README.md">Anand Kumar(Discord-White Devil#4648)</a> | 
-						 <span class="far fa-copyright"></span> 2021 All rights reserved.</span> </footer>
-  
-						 
-						 
-						 
-						 
-						 
 
-		
-				
+
+
+						<span class="typing"></span>
+					</div>
+				</div>
+			</div>
+	</section>
+
+
+
+	<footer class="footer">
+		<span>Created By <i class="fab fa-github"></i> <a
+				href="https://github.com/Whitedevil4648/INIT-2022-COVID-SITE/blob/main/README.md">Anand
+				Kumar(Discord-White Devil#4648)</a> |
+			<span class="far fa-copyright"></span> 2021 All rights reserved.</span>
+	</footer>
+
+
+
+
+
+
+
+
+
 </body>
+
 </html>
